Refactor theme normalization into shared module
